# management/registry/openapi/openapi.yaml
basePath: /
definitions:
...
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  handler.NodeRegisterRequest:
    properties:
      cf_url:
        example: https://tunnel.example.com
        type: string
      hostname:
        example: node-01
        type: string
      mac:
        example: AA:BB:CC:DD:EE:FF
        type: string
      mac6:
        example: AABBCC
        type: string
      mesh_ip:
        example: 100.96.0.10
        type: string
      mesh_tunnel_id:
        example: 704320e0-f87b-4ca8-a45e-481c1fcd7462
        type: string
      tunnel_id:
        example: abc123
        type: string
      xray_uuid:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  handler.NodeRegisterResponse:
    properties:
      hostname:
        example: node-01
        type: string
      registered_at:
        example: "2025-01-01 00:00:00"
        type: string
      status:
        example: registered
        type: string
    type: object
  handler.NodeUpdateRequest:
    properties:
      easytier_ip:
        example: 10.0.0.1
        type: string
      enabled:
        example: true
        type: boolean
      frp_port:
        example: 7000
        type: integer
      location:
        example: Tokyo
        type: string
      mesh_ip:
        example: 100.96.0.10
        type: string
      mesh_tunnel_id:
        example: 704320e0-f87b-4ca8-a45e-481c1fcd7462
        type: string
      note:
        example: primary node
        type: string
      region:
        example: asia
        type: string
      status:
        example: online
        type: string
      tailscale_ip:
        example: 100.64.0.1
        type: string
      weight:
        example: 10
        type: integer
    type: object
  handler.NodeXrayUsersResponse:
    properties:
      mac:
        type: string
      users:
        items:
          $ref: '#/definitions/handler.XrayUser'
        type: array
    type: object
...
  handler.StatusResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
  handler.XrayUser:
    properties:
      email:
        type: string
      uuid:
        type: string
    type: object
...
  model.Node:
    properties:
      cf_url:
        type: string
      easytier_ip:
        type: string
      enabled:
        type: boolean
      frp_port:
        type: integer
      hostname:
        type: string
      last_seen:
        type: string
      location:
        type: string
      mac:
        type: string
      mac6:
        type: string
      mesh_ip:
        type: string
      mesh_tunnel_id:
        type: string
      note:
        type: string
      probe_status:
        type: string
      region:
        type: string
      registered_at:
        type: string
      status:
        type: string
      tailscale_ip:
        type: string
      tunnel_id:
        type: string
      weight:
        type: integer
      xray_uuid:
        type: string
    type: object
  model.NodeStatusCheck:
    properties:
      checked_at:
        type: string
      cpu_pct:
        type: number
      disk_pct:
        type: number
      hostname:
        type: string
      latency_ms:
        type: integer
      location:
        type: string
      mac:
        type: string
      mem_pct:
        type: number
      status:
        type: string
      uptime_sec:
        type: integer
    type: object
...
info:
  contact: {}
  description: Hive management plane API for nodes and subscriptions.
  title: Hive Node Registry API
  version: 0.1.0
paths:
...
  /admin/node-status:
    get:
      description: 返回所有节点的最新探测状态（CPU、内存、磁盘、延迟等）
      operationId: AdminNodeStatus
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.NodeStatusCheck'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AdminSessionCookie: []
      summary: 获取节点状态列表
      tags:
      - admin
...
  /nodes:
    get:
      description: List all nodes with optional filters
      operationId: NodesList
      parameters:
      - description: filter by status
        in: query
        name: status
        type: string
      - description: filter by enabled (0 or 1)
        in: query
        name: enabled
        type: string
      - description: filter by region
        in: query
        name: region
        type: string
      - description: search hostname, location, note, or mac
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Node'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AdminSessionCookie: []
      summary: List nodes
      tags:
      - admin
  /nodes/{mac}:
    delete:
      description: Delete a node by MAC address
      operationId: NodeDelete
      parameters:
      - description: node MAC address
        in: path
        name: mac
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AdminSessionCookie: []
      summary: Delete a node
      tags:
      - admin
    get:
      description: Get a single node by MAC address
      operationId: NodeGet
      parameters:
      - description: node MAC address
        in: path
        name: mac
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Node'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AdminSessionCookie: []
      summary: Get a node
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: Update allowed fields of a node by MAC address
      operationId: NodeUpdate
      parameters:
      - description: node MAC address
        in: path
        name: mac
        required: true
        type: string
      - description: fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.NodeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AdminSessionCookie: []
      summary: Update a node
      tags:
      - admin
  /nodes/{mac}/xray-users:
    get:
      description: |-
        Returns the set of VLESS clients (uuid + email) that should be active
        on this node's Xray inbound: all valid (active, non-expired, under-quota)
        customer subscriptions whose plan grants access to a line containing this node.
      operationId: NodeXrayUsers
      parameters:
      - description: node MAC
        in: path
        name: mac
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.NodeXrayUsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Xray users for a node
      tags:
      - nodes
...
  /nodes/register:
    post:
      consumes:
      - application/json
      description: Register a new node or update an existing one by MAC address
      operationId: NodeRegister
      parameters:
      - description: node info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.NodeRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.NodeRegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a node
      tags:
      - nodes
...
schemes:
- https
- http
securityDefinitions:
  AdminSessionCookie:
    description: '"HttpOnly session cookie for admin UI"'
    in: cookie
    name: hive_admin_session
    type: apiKey
  BearerAuth:
    description: '"Authorization: Bearer <API_SECRET>"'
    in: header
    name: Authorization
    type: apiKey
  CustomerSessionCookie:
    description: '"HttpOnly session cookie for customer portal"'
    in: cookie
    name: hive_customer_session
    type: apiKey
swagger: "2.0"